// src/errors/database/db-error.enum.ts
export enum DBErrorCode {
  // 23xxx
  DUPLICATE_RESOURCE = "DUPLICATE_RESOURCE", // 23505
  FOREIGN_KEY_VIOLATION = "FOREIGN_KEY_VIOLATION", // 23503
  NOT_NULL_VIOLATION = "NOT_NULL_VIOLATION", // 23502
  CHECK_CONSTRAINT_VIOLATION = "CHECK_CONSTRAINT_VIOLATION", // 23514
  INTEGRITY_CONSTRAINT_VIOLATION = "INTEGRITY_CONSTRAINT_VIOLATION",

  // 22xxx
  INVALID_INPUT_SYNTAX = "INVALID_INPUT_SYNTAX", // 22P02
  DATA_EXCEPTION = "DATA_EXCEPTION",

  // 40xxx
  SERIALIZATION_FAILURE = "SERIALIZATION_FAILURE", // 40001
  DEADLOCK_DETECTED = "DEADLOCK_DETECTED", // 40P01
  TRANSACTION_ROLLBACK = "TRANSACTION_ROLLBACK",

  // 28xxx
  INVALID_AUTHORIZATION = "INVALID_AUTHORIZATION",

  // 42xxx
  SYNTAX_ERROR = "SYNTAX_ERROR",

  // 53xxx
  INSUFFICIENT_RESOURCES = "INSUFFICIENT_RESOURCES",

  // 57xxx
  OPERATOR_INTERVENTION = "OPERATOR_INTERVENTION",

  // 58xxx
  SYSTEM_ERROR = "SYSTEM_ERROR",

  // custom mapping
  RESOURCE_NOT_FOUND = "RESOURCE_NOT_FOUND",
  DATABASE_ERROR = "DATABASE_ERROR",
}